Add an example TileUpdateTransformer
