+++ lib/flutter_map.dart
@@ -40,6 +40,8 @@ export 'package:flutter_map/src/layer/tile_layer/tile_provider/file_tile_provide
     if (dart.library.html) 'package:flutter_map/src/layer/tile_layer/tile_provider/file_tile_provider_web.dart';
 export 'package:flutter_map/src/layer/tile_layer/tile_provider/tile_provider_io.dart'
     if (dart.library.html) 'package:flutter_map/src/layer/tile_layer/tile_provider/tile_provider_web.dart';
+export 'package:flutter_map/src/layer/tile_layer/tile_update_event.dart';
+export 'package:flutter_map/src/layer/tile_layer/tile_update_transformer.dart';
 
 /// Renders a map composed of a list of layers powered by [LayerOptions].
 ///
